Fix: modal shows info about error occured while removing employee
